refactor(react-localization): host the canonical resolveLabel helper
resolveLabel (display-key → translated label, last-segment fallback) lived in
react-ui-shell-admin/workspace-utils but is a generic i18n helper also needed by
manifest-driven entity UI. Move it to @granit/react-localization and have
shell-admin re-export it from workspace-utils, so existing consumers are
unaffected. Lets the upcoming @granit/react-ui-entities avoid a layer-inversion
dependency on shell-admin.
